fixed tests, moved reporters into own sub-dir
