+++ lib/cli.js
@@ -73,11 +73,11 @@ module.exports = {
         }
 
         if (options["--jslint-reporter"]) {
-            customReporter = "./jslint_reporter.js";
+            customReporter = "./reporters/jslint_xml.js";
         }
 
         if (options["--show-non-errors"]) {
-            customReporter = "./non_error_reporter.js";
+            customReporter = "./reporters/non_error.js";
         }
 
         if (customConfig) {
