+++ lib/hint.js
@@ -2,6 +2,7 @@ var _fs = require('fs'),
     _sys = require('sys'),
     _path = require('path'),
     _jshint = require('./../packages/jshint/jshint.js'),
+    _reporter = require('./reporters/default'),
     _config;
 
 function _lint(file, results, data) {
@@ -24,8 +25,11 @@ function _lint(file, results, data) {
     }
 
     lintdata = _jshint.JSHINT.data();
-    lintdata.file = file;
-    data.push(lintdata);
+
+    if (lintdata) {
+        lintdata.file = file;
+        data.push(lintdata);
+    }
 }
 
 function _collect(path, files) {
@@ -38,29 +42,12 @@ function _collect(path, files) {
     }
 }
 
-function _reporter(results) {
-    var len = results.length,
-        str = '',
-        file, error;
-
-    results.forEach(function (result) {
-        file = result.file;
-        error = result.error;
-        str += file  + ': line ' + error.line + ', col ' +
-            error.character + ', ' + error.reason + '\n';
-    });
-
-    _sys.puts(len > 0 ? (str + "\n" + len + ' error' + ((len === 1) ? '' : 's')) : "Lint Free!");
-    process.exit(len > 0 ? 1 : 0);
-}
-
 module.exports = {
     hint: function (targets, config, reporter) {
         var files = [],
             results = [],
             data = [];
 
-        if (!reporter) reporter = _reporter;
         _config = config || null;
 
         targets.forEach(function (target) {
@@ -71,6 +58,6 @@ module.exports = {
             _lint(file, results, data);
         });
 
-        reporter(results, data);
+        (reporter || _reporter)(results, data);
     }
 };
