+++ lib/reporters/jslint_xml.js
