+++ lib/reporters/non_error.js
